Add files via upload
A few stl files for prints I've designed to make things a bit easier.

Corner guides to allow setting the build plate down and keep it aligned accurately.  I've not seen any indications that there's any chance of collision between these guides and the toolhead but I make no guarantee.  Use at your own risk.

The probe storage is just to have a handy place in the printer itself to store that probe. Can be attached to the 2020 extrusions in any space where there's no risk of collision with build plate or toolhead.  I plan to place it on the bottom right that space seems like there's zero chance of any movement of build plate or toolhead conflicting.

The button holder is just a handy tool to put on your finger to make holding the button on the toolhead down while putting it into DFU mode for flashing.